Move admin rest api methods to own module.
Separate out the admin rest api functions into a submodule of admin.
This will allow the entirety of the admin rest features to be guarded by
a rust feature.
